Refactor - Update budget and reveneu parameters to Long.

+++ app/src/main/java/com/github/androidpirate/capsulereviews/data/response/movie/MovieResponse.kt
@@ -7,7 +7,7 @@ data class MovieResponse(
     @SerializedName("backdrop_path")
     val backdropPath: String,
     @SerializedName("budget")
-    val budget: Int,
+    val budget: Long,
     @SerializedName("genres")
     val genres: List<Genre>,
     @SerializedName("id")
@@ -27,7 +27,7 @@ data class MovieResponse(
     @SerializedName("release_date")
     val releaseDate: String,
     @SerializedName("revenue")
-    val revenue: Int,
+    val revenue: Long,
     @SerializedName("runtime")
     val runtime: Int,
     @SerializedName("status")
